Remove ftag for now.
ftag's purpose is clear, but still unsure how it can be made
portable without indexing the entire home directory. If tags are
attached to a file's identity, ie. a checksum, then the program
is required to index and read every file; this has to be done
on every home directory mutation, even with a heuristic it's way
too much work for a simple 'script'.